refactor: split large commands file into modules
