+++ src/commands/mod.rs
@@ -6,12 +6,9 @@ mod replay;
 mod run;
 
 // Re-export only the functions that execute commands
-pub use data::data_description::DataStatus;
-pub use data::fetch::FetchCommand;
-pub use data::{clean_data, package_data_to_archive};
-
-pub use new::create_kerblam_project;
-pub use other::ignore;
-pub use package::package_pipe;
-pub use replay::replay;
-pub use run::kerblam_run_project;
+pub use data::DataCommand;
+pub use new::NewCommand;
+pub use other::IgnoreCommand;
+pub use package::PackageCommand;
+pub use replay::ReplayCommand;
+pub use run::RunCommand;
